add many small fixes
Fixed display of modal on mobile version.
Added modal window scroll if it goes beyond the screen border. Fixed a bug with a variable in the modal window when clicking on the "add to watch" button. Added test check for login. Added correct display of genres in the modal window.